+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Update: Bitcoin Core v25.1 [details](https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-25.1.md)
 - Update: LND v0.17.2-beta [details](https://github.com/lightningnetwork/lnd/releases/tag/v0.17.2-beta)
 - Update: BTCPayServer v1.11.7 [details](https://github.com/btcpayserver/btcpayserver/releases/tag/v1.11.7)
+- Update: LNbits 0.11.1 [details](https://github.com/lnbits/lnbits/releases/tag/0.11.1)
 
 ## What's new in Version 1.10.0 of RaspiBlitz?